+++ WithSecure/withsecure/models.py
@@ -11,3 +11,7 @@ class WithSecureModuleConfiguration(BaseModel):
 class RemoteOperationResponse(BaseModel):
     multistatus: list[Any]
     transactionId: str
+
+
+class ResponseActionResponse(BaseModel):
+    id: str | None = None